# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Xuthority API",
    "version": "1.0.0",
    "description": "Automatically generated API documentation.",
    "contact": {
      "name": "Xuthority",
      "url": "https://xuthority.com",
      "email": "support@xuthority.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://xuthority.com/terms",
    "xuthority": "https://xuthority.com",
    "xuthorityDev": "https://xuthority-dev.com",
    "xuthorityDevApi": "http://localhost:8081/api/v1",
    "xuthorityDevApiDocs": "http://localhost:8081/api-docs",
    "xuthorityDevApiDocsSwagger": "http://localhost:8081/swagger.json",
    "xuthorityDevApiDocsSwaggerUi": "http://localhost:8081/api-docs",
    "xuthorityDevApiDocsSwaggerUiSwagger": "http://localhost:8081/swagger.json"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Blog": {
        "type": "object",
        "required": [
          "authorName",
          "title",
          "description",
          "tag",
          "resourceCategoryId"
        ],
        "properties": {
          "authorName": {
            "type": "string",
            "maxLength": 100
          },
          "designation": {
            "type": "string",
            "maxLength": 100
          },
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 200
          },
          "slug": {
            "type": "string"
          },
          "mediaUrl": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "minLength": 10
          },
          "watchUrl": {
            "type": "string"
          },
          "tag": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "resourceCategoryId": {
            "type": "string",
            "format": "objectId"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/blogs": {
      "post": {
        "summary": "Create a new blog",
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Blog"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get all blogs",
        "tags": [
          "Blogs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blogs retrieved successfully"
          }
        }
      }
    },
...
    "/api/v1/blogs/category/{categoryId}": {
      "get": {
        "summary": "Get blogs by category",
        "tags": [
          "Blogs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blogs retrieved successfully"
          }
        }
      }
    },
    "/api/v1/blogs/tag/{tag}": {
      "get": {
        "summary": "Get blogs by tag",
        "tags": [
          "Blogs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tag",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blogs retrieved successfully"
          }
        }
      }
    },
    "/api/v1/blogs/slug/{slug}": {
      "get": {
        "summary": "Get blog by slug",
        "tags": [
          "Blogs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog retrieved successfully"
          },
          "404": {
            "description": "Blog not found"
          }
        }
      }
    },
    "/api/v1/blogs/{id}": {
      "get": {
        "summary": "Get blog by ID",
        "tags": [
          "Blogs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog retrieved successfully"
          },
          "404": {
            "description": "Blog not found"
          }
        }
      },
      "put": {
        "summary": "Update blog",
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Blog"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog updated successfully"
          },
          "404": {
            "description": "Blog not found"
          }
        }
      },
      "delete": {
        "summary": "Delete blog",
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog deleted successfully"
          },
          "404": {
            "description": "Blog not found"
          }
        }
      }
    },
    "/api/v1/blogs/{id}/toggle-status": {
      "patch": {
        "summary": "Toggle blog status",
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog status toggled successfully"
          },
          "404": {
            "description": "Blog not found"
          }
        }
      }
    },
...
  },
...
}